REF: Move Cloud Run deployment module and Dockerfile into subdirectory
